# api/docs/admin/admin_swagger.yaml
basePath: /api
definitions:
...
  admin.PeerBatchDeleteForm:
    properties:
      row_ids:
        items:
          type: integer
        type: array
    required:
    - row_ids
    type: object
  admin.PeerForm:
    properties:
      alias:
        type: string
      cpu:
        type: string
      group_id:
        type: integer
      hostname:
        type: string
      id:
        type: string
      memory:
        type: string
      os:
        type: string
      row_id:
        type: integer
      username:
        type: string
      uuid:
        type: string
      version:
        type: string
    type: object
...
  admin.SimpleDataQuery:
    properties:
      ids:
        items:
          type: string
        type: array
    type: object
...
  model.AddressBookCollectionRule:
    properties:
      collection_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      rule:
        description: '0:  1:  2:   3:'
        maximum: 3
        minimum: 1
        type: integer
      to_id:
        type: integer
      type:
        description: '1:  2:'
        maximum: 2
        minimum: 1
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    required:
    - collection_id
    - rule
    - to_id
    - type
    type: object
  model.AddressBookCollectionRuleList:
    properties:
      list:
        items:
          $ref: '#/definitions/model.AddressBookCollectionRule'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  model.Peer:
    properties:
      alias:
        type: string
      cpu:
        type: string
      created_at:
        type: string
      group_id:
        type: integer
      hostname:
        type: string
      id:
        type: string
      last_online_ip:
        type: string
      last_online_time:
        type: integer
      memory:
        type: string
      os:
        type: string
      row_id:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/model.User'
      user_id:
        type: integer
      username:
        type: string
      uuid:
        type: string
      version:
        type: string
    type: object
  model.PeerList:
    properties:
      list:
        items:
          $ref: '#/definitions/model.Peer'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  model.StatusCode:
    enum:
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - COMMON_STATUS_ENABLE
    - COMMON_STATUS_DISABLED
...
  model.User:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      group_id:
        type: integer
      id:
        type: integer
      is_admin:
        type: boolean
      nickname:
        type: string
      remark:
        type: string
      status:
        $ref: '#/definitions/model.StatusCode'
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
info:
  contact: {}
  title: API
  version: "1.0"
paths:
...
  /admin/my/address_book_collection_rule/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Personal address book collection rule form payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AddressBookCollectionRule'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      tags:
      - ""
  /admin/my/address_book_collection_rule/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: Personal address book collection rule deletion form payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AddressBookCollectionRule'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      tags:
      - ""
  /admin/my/address_book_collection_rule/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page number for personal address book collection rules
        in: query
        name: page
        type: integer
      - description: Number of personal collection rules per page
        in: query
        name: page_size
        type: integer
      - description: Filter for rules owned by the current user
        in: query
        name: is_my
        type: integer
      - description: id
        in: query
        name: user_id
        type: integer
      - description: id
        in: query
        name: collection_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AddressBookCollectionRuleList'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      tags:
      - ""
  /admin/my/address_book_collection_rule/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: Personal address book collection rule update form payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AddressBookCollectionRule'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      tags:
      - ""
...
  /admin/peer/batchDelete:
    post:
      consumes:
      - application/json
      parameters:
      - description: id
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.PeerBatchDeleteForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      tags:
      - ""
  /admin/peer/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Peer form payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.PeerForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Peer'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      tags:
      - ""
  /admin/peer/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: Peer deletion form payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.PeerForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      tags:
      - ""
  /admin/peer/detail/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Peer'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      tags:
      - ""
  /admin/peer/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page number for the peer list
        in: query
        name: page
        type: integer
      - description: Number of peers per page
        in: query
        name: page_size
        type: integer
      - description: Last-online time offset filter
        in: query
        name: time_ago
        type: integer
      - description: ID
        in: query
        name: id
        type: string
      - description: Hostname filter
        in: query
        name: hostname
        type: string
      - description: 'uuids '
        in: query
        name: uuids
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.PeerList'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      tags:
      - ""
  /admin/peer/simpleData:
    post:
      consumes:
      - application/json
      description: Authenticated peer version lookup used by the admin panel.
      parameters:
      - description: Peer ID lookup payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.SimpleDataQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.PeerList'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: Get peer version data
      tags:
      - Peer
  /admin/peer/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: Peer update form payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.PeerForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Peer'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      tags:
      - ""
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  token:
    in: header
    name: api-token
    type: apiKey
swagger: "2.0"